start separating fock things and physics
